Clean up list select items to make using a DragOverlay easier
